requirements.txt
# requirements.txt

# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0

# Google Cloud
google-cloud-aiplatform==1.74.0
google-cloud-firestore==2.19.0
google-genai==1.1.0

# MCP
mcp==1.3.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
httpx==0.28.1
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0